@@ -1,34 +1,81 @@
 import math
+from typing import Union
+
+
+class SchedulerMismatchError(Exception):
+    """
+    Exception for indicating a mismatch in the number of expected
+    documents versus the total number of actually uploaded documents
+    to the target database
+    """
+
+    def __init__(self, completed_documents: int, expected_documents: int):
+
+        self.completed_documents = completed_documents
+        self.expected_documents = expected_documents
+
+        message = (
+            f"Difference found between the number of completed documents [{self.completed_documents}] "
+            f"in the indexing process and the number of expected documents [{self.expected_documents}] "
+            "to be indexed based off the collection size. "
+            "This error can occur if the documents uploaded don't have an _id field with MongoDB. "
+            "Please verify the document structure at the upload phase"
+        )
+        super().__init__(message)
 
 
 class Schedule:
-    def __init__(self, total, batch_size):
+    def __init__(self, total: int, batch_size: int):
         self._batch_size = batch_size
-        self._state = ""
-
         self.total = total
+
+        self._state = "initialization"
         self.scheduled = 0
         self.finished = 0
 
+    @classmethod
+    def _calculate_percentage(
+        self, numerator: Union[int, float], denominator: Union[int, float], ceiling: bool
+    ) -> float:
+        """
+        Calculate the percentage given two numbers and return that value
+
+        -> numerator / denominator
+
+        if ceiling is applied then we apply the ceiling function to the result
+        -> ceiling[numerator / denominator]
+        """
+        try:
+            percentage = numerator / denominator
+        except ZeroDivisionError as div_error:
+            raise div_error
+
+        if ceiling:
+            percentage = math.ceil(percentage)
+        return percentage
+
     @property
     def _batch(self):
-        return math.ceil(self.scheduled / self._batch_size)
+        return self._calculate_percentage(self.scheduled, self._batch_size, True)
 
     @property
     def _batches(self):
-        return math.ceil(self.total / self._batch_size)
+        return self._calculate_percentage(self.total, self._batch_size, True)
 
     @property
     def _percentage(self):
-        _percentage = self.scheduled / self.total * 100
+        _percentage = self._calculate_percentage(self.scheduled, self.total, False) * 100
         return "%.1f%%" % _percentage
 
-    def suffix(self, string):
-        return " ".join((string, "#%d/%d %s" % (self._batch, self._batches, self._percentage)))
+    def suffix(self, suffix_value: str) -> str:
+        batch_repr = f"#{self._batch}/{self._batches} {self._percentage}"
+        suffix_repr = f"{suffix_value} {batch_repr}"
+        return suffix_repr
 
     def completed(self):
-        if self.finished != self.total:
-            raise ValueError(self.finished, self.total)
+        if not self.finished == self.total:
+            raise SchedulerMismatchError(self.finished, self.total)
+        self._state = "done"
 
     def __iter__(self):
         return self
@@ -37,51 +84,17 @@ class Schedule:
         if self.scheduled >= self.total:
             self._state = "pending, waiting for completion,"
             raise StopIteration()
+
         self.scheduled += self._batch_size
-        if self.scheduled > self.total:
-            self.scheduled = self.total
+        self.scheduled = min(self.scheduled, self.total)
         self._state = self.suffix("running, on batch") + ","
         return self._batch
 
-    def __str__(self):
-        return " ".join(
-            f"""
-            <Schedule {"done" if self.finished >= self.total else self._state}
-                total={self.total} scheduled={self.scheduled} finished={self.finished}>
-            """.split()
+    def __str__(self) -> str:
+        schedule_str = (
+            f"<schedule {self._state} "
+            f"total={self.total} "
+            f"scheduled={self.scheduled} "
+            f"finished={self.finished}>"
         )
-
-
-def test_01():
-    schedule = Schedule(100, 10)
-    for batch in schedule:
-        print(batch)
-        print(schedule)
-
-
-def test_02():
-    schedule = Schedule(25, 10)
-    for batch in schedule:
-        print(batch)
-        print(schedule)
-        print(schedule.suffix("Task"))
-
-
-def test_03():
-    schedule = Schedule(0, 10)
-    for batch in schedule:
-        print(batch)
-        print(schedule)
-        print(schedule.suffix("Task"))
-
-
-def test_04():
-    schedule = Schedule(1, 10)
-    for batch in schedule:
-        print(batch)
-        print(schedule)
-        print(schedule.suffix("Task"))
-
-
-if __name__ == "__main__":
-    test_02()
+        return schedule_str
