@@ -2,6 +2,7 @@
 Utility functions for parsing flatfiles,
 mapping to JSON, cleaning.
 """
+
 import csv
 
 # see tabfile_feeder(coerce_unicode) if needed
@@ -250,53 +251,71 @@ def rec_handler(infile, block_end="\n", skip=0, include_block_end=False, as_list
 # ===============================================================================
 
 
-# if dict value is a list of length 1, unlist
 def unlist(d):
+    """Recursively unlist elements in a dictionary.
+
+    If an element in a list is a dictionary, apply the unlist function to it recursively.
+    If a list contains only one element, replace the list with that single element.
+    The function operates on all keys within the dictionary.
+
+    :param d: a dictionary to unlist
+    :return: the modified dictionary with unlisted elements
+    """
     for key, val in d.items():
         if isinstance(val, list):
-            if len(val) == 1:
-                d[key] = val[0]
+            d[key] = [unlist(v) if isinstance(v, dict) else v for v in val]
+            if len(d[key]) == 1:
+                d[key] = d[key][0]
         elif isinstance(val, dict):
-            unlist(val)
+            d[key] = unlist(val)
+    return d
+
+# Internal helper function called by unlist_incexcl
+def _should_unlist(path, include_keys=None, exclude_keys=None):
+    if include_keys is not None and exclude_keys is not None:
+        return path in include_keys and path not in exclude_keys
+    elif include_keys is not None:
+        return path in include_keys
+    elif exclude_keys is not None:
+        return path not in exclude_keys
+    return True
+
+# Internal helper function called by unlist_incexcl
+def _unlist_helper(d, include_keys=None, exclude_keys=None, keys=None):
+    keys = keys or []
+    if isinstance(d, dict):
+        for key, val in list(d.items()):
+            path = ".".join(keys + [key])
+            if isinstance(val, list):
+                if len(val) == 1 and _should_unlist(path, include_keys, exclude_keys):
+                    d[key] = _unlist_helper(
+                        val[0], include_keys, exclude_keys, keys + [key])
+                else:
+                    d[key] = [_unlist_helper(
+                        item, include_keys, exclude_keys, keys + [key]) for item in val]
+            elif isinstance(val, dict):
+                d[key] = _unlist_helper(
+                    val, include_keys, exclude_keys, keys + [key])
     return d
 
 
 def unlist_incexcl(d, include_keys=None, exclude_keys=None):
     """Unlist elements in a document.
 
-    If there is 1 value in the list, set the element to that value.  Otherwise,
+    If there is 1 value in the list, set the element to that value. Otherwise,
     leave the list unchanged.
 
-    By default, traverse all keys
-    If include_keys is specified, only traverse the list from include_keys a.b, a.b.c
-    If exclude_keys is specified, only exclude the list from exclude_keys
+    By default, traverse all keys.
+    If include_keys is specified, only traverse the list for keys in include_keys.
+    If exclude_keys is specified, exclude the list for keys in exclude_keys.
 
     :param d: a dictionary to unlist
     :param include_keys: only unlist these keys (optional)
     :param exclude_keys: exclude all other keys except these keys (optional)
-    :return: generate key, value pairs
+    :return: the modified dictionary
     """
-
-    def unlist_helper(d, include_keys=None, exclude_keys=None, keys=None):
-        include_keys = include_keys or []
-        exclude_keys = exclude_keys or []
-        keys = keys or []
-        if isinstance(d, dict):
-            for key, val in d.items():
-                if isinstance(val, list):
-                    if len(val) == 1:
-                        path = ".".join(keys + [key])
-                        if include_keys:
-                            if path in include_keys:
-                                d[key] = val[0]
-                        elif path not in exclude_keys:
-                            d[key] = val[0]
-                elif isinstance(val, dict):
-                    unlist_helper(val, include_keys, exclude_keys, keys + [key])
-
-    unlist_helper(d, include_keys, exclude_keys, [])
-    return d
-
+    return _unlist_helper(d, include_keys, exclude_keys)
+    
 
 def list_split(d, sep):
     """Split fields by sep into comma separated lists, strip."""
@@ -942,6 +961,10 @@ def merge_struct(v1, v2, aslistofdict=None, include=None, exclude=None):
     :param include: when given a list of strings, only merge these keys (optional)
     :param exclude: when given a list of strings, exclude these keys from merging (optional)
     """
+
+    # https://docs.python.org/3/library/stdtypes.html
+    supported_primitive_types = (int, float, complex, str, bool, type(None))
+
     if isinstance(v1, list):
         if isinstance(v2, list):
             v1 = v1 + [x for x in v2 if x not in v1]
@@ -984,8 +1007,8 @@ def merge_struct(v1, v2, aslistofdict=None, include=None, exclude=None):
             else:
                 v1[k] = v2[k]
 
-    elif isinstance(v1, str) or isinstance(v1, int) or isinstance(v1, float):
-        if isinstance(v2, str) or isinstance(v2, int) or isinstance(v2, float):
+    elif isinstance(v1, supported_primitive_types):
+        if isinstance(v2, supported_primitive_types):
             if v1 != v2:
                 v1 = [v1, v2]
             else:
@@ -993,7 +1016,8 @@ def merge_struct(v1, v2, aslistofdict=None, include=None, exclude=None):
         else:
             return merge_struct(v2, v1)
     else:
-        raise TypeError("dunno how to merge type %s" % type(v1))
+        error_message = f"Unknown type {type(v1)} while attempting to merge structures"
+        raise TypeError(error_message)
 
     return v1
 
