@@ -99,12 +99,10 @@ class _UserString(UserString):
         return f"{type(self).__name__}({self.data})"
 
 
-class TemplateStr(_UserString):
-    ...
+class TemplateStr(_UserString): ...
 
 
-class RenderedStr(_UserString):
-    ...
+class RenderedStr(_UserString): ...
 
 
 class RepositoryConfig(UserDict):
@@ -162,12 +160,10 @@ class _SnapshotResult(UserDict):
         return f"{type(self).__name__}({str(self.data)})"
 
 
-class CumulativeResult(_SnapshotResult):
-    ...
+class CumulativeResult(_SnapshotResult): ...
 
 
-class StepResult(_SnapshotResult):
-    ...
+class StepResult(_SnapshotResult): ...
 
 
 class SnapshotEnv:
@@ -185,8 +181,7 @@ class SnapshotEnv:
         self.wtime = kwargs.get("monitor_delay", 15)
 
     def _doc(self, index):
-        doc = get_src_build().find_one(
-            {f"index.{index}.environment": self.idxenv})
+        doc = get_src_build().find_one({f"index.{index}.environment": self.idxenv})
         if not doc:  # not asso. with a build
             raise ValueError("Not a hub-managed index.")
         return doc  # TODO UNIQUENESS
@@ -194,8 +189,7 @@ class SnapshotEnv:
     def setup_log(self, index):
         build_doc = self._doc(index)
         log_name = build_doc["target_name"] or build_doc["_id"]
-        log_folder = os.path.join(
-            btconfig.LOG_FOLDER, "build", log_name, "snapshot") if btconfig.LOG_FOLDER else None
+        log_folder = os.path.join(btconfig.LOG_FOLDER, "build", log_name, "snapshot") if btconfig.LOG_FOLDER else None
         self.logger, _ = get_logger(index, log_folder=log_folder, force=True)
 
     def snapshot(self, index, snapshot=None, recreate_repo=False):
@@ -213,8 +207,7 @@ class SnapshotEnv:
 
                 job = await self.job_manager.defer_to_thread(
                     self.pinfo.get_pinfo(step, snapshot),
-                    partial(getattr(self, state.func), cfg, index,
-                            snapshot, recreate_repo=recreate_repo),
+                    partial(getattr(self, state.func), cfg, index, snapshot, recreate_repo=recreate_repo),
                 )
                 try:
                     dx = await job
@@ -259,8 +252,7 @@ class SnapshotEnv:
         try:
             repo.verify(config=cfg)
         except TransportError as tex:
-            raise RepositoryVerificationFailed(
-                {"error": tex.error, "detail": tex.info["error"]})
+            raise RepositoryVerificationFailed({"error": tex.error, "detail": tex.info["error"]})
 
         return {
             "__REPLACE__": True,
@@ -516,14 +508,12 @@ class SnapshotManager(BaseManager):
         jobs = []
         try:
             for environment, snapshot_names in snapshots_data.items():
-                job = self.job_manager.submit(
-                    partial(delete, environment, snapshot_names))
+                job = self.job_manager.submit(partial(delete, environment, snapshot_names))
                 jobs.append(job)
             tasks = asyncio.gather(*jobs)
             tasks.add_done_callback(done)
         except Exception as ex:
-            logging.exception(
-                "Error while deleting snapshots. error: %s", ex, extra={"notify": True})
+            logging.exception("Error while deleting snapshots. error: %s", ex, extra={"notify": True})
         return jobs
 
     def delete_snapshot_from_db(self, build_name, snapshot_name):
