@@ -1,8 +1,13 @@
+from functools import wraps
+from typing import Any, Iterable, Optional, Union, cast
 import collections
 import json
 import logging
 import os
 import sqlite3
+
+from pymongo import InsertOne, ReplaceOne, UpdateOne
+from typing import Any, Mapping, TypeVar
 from functools import wraps
 
 from biothings.utils.common import find_value_in_doc, json_serial
@@ -14,6 +19,8 @@ from biothings.utils.serializer import json_loads
 config = None
 
 logger = logging.getLogger(__name__)
+
+DocumentType = TypeVar("DocumentType", bound=Mapping[str, Any])
 
 
 def requires_config(func):
@@ -183,6 +190,86 @@ class DatabaseClient(IDatabase):
         return Database(self.sqlite_db_folder, name)
 
 
+class Sqlite3BulkWriteResult:
+    """
+    An object wrapper for bulk API write results
+
+    This mimics the structure defined by pymongo defined
+    by `pymongo.results.BulkWriteResult`
+    """
+
+    __slots__ = ("__bulk_api_result",)
+
+    def __init__(self, bulk_api_result: Iterable[Any]):
+        self.__bulk_api_result = bulk_api_result
+
+    def __repr__(self) -> str:
+        return f"{self.__class__.__name__}({self.__bulk_api_result!r}"
+
+    @property
+    def bulk_api_result(self) -> Iterable[Any]:
+        """
+        The raw bulk API result
+        """
+        return self.__bulk_api_result
+
+    @property
+    def inserted_count(self) -> int:
+        """
+        The number of documents inserted
+        """
+        return len(self.__bulk_api_result)
+
+    @property
+    def matched_count(self) -> int:
+        """
+        The number of documents matched for an update
+        """
+        logger.warning("sqlite3 has no concept of a match counted")
+        return 0
+
+    @property
+    def modified_count(self) -> int:
+        """
+        The number of documents modified
+        """
+        return cast(int, self.__bulk_api_result.get("nModified"))
+
+    @property
+    def deleted_count(self) -> int:
+        """
+        The number of documents deleted
+        """
+        return cast(int, self.__bulk_api_result.get("nRemoved"))
+
+    @property
+    def upserted_count(self) -> int:
+        """
+        The number of documents upserted
+        """
+        logger.warning("sqlite3 has no concept of a upserted counted")
+        return 0
+
+    @property
+    def upserted_ids(self) -> Optional[dict[int, Any]]:
+        """
+        A map of operation index to the _id of the upserted document
+        """
+        if self.__bulk_api_result:
+            return {upsert["index"]: upsert["_id"] for upsert in self.bulk_api_result["upserted"]}
+        return None
+
+
+class Sqlite3BulkWriteError(Exception):
+    """
+    Bulk write error specifically targetting the sqlite3
+    database backend. We won't get any metrics on the failed insertion
+    other than the table and the column, we'll have to do a bit more processing
+    on the documents that were attempted to be written. Thus we have a Exception
+    that is essentially a wrapper around sqlite3.IntegrityError
+    """
+
+
 class Collection(object):
     def __init__(self, colname, db):
         self.colname = colname
@@ -336,6 +423,28 @@ class Collection(object):
         else:
             raise NotImplementedError("find: args=%s kwargs=%s" % (repr(args), repr(kwargs)))
 
+    def id_search(self, id_collection: list[str]) -> list[tuple]:
+        """
+        Method for bulk searching against the _id column in the database.
+        Primarily used for determining the culprit to the uniqueness integrity violation
+        from a bulk write.
+
+        args:
+            id_collection: list of strings representing the document _id value
+
+        output:
+        """
+        discovered_id = []
+        with self.get_conn() as conn:
+            for id_value in id_collection:
+                try:
+                    id_verification_query = f"SELECT _id, document FROM {self.colname} WHERE _id = '{id_value}'"
+                    discovered_id.extend(conn.execute(id_verification_query).fetchall())
+                except Exception as gen_exp:
+                    logger.exception(gen_exp)
+                    logger.warning("Skipping %s for id search", id_value)
+        return discovered_id
+
     def insert_one(self, doc: dict, *args, **kwargs) -> None:
         """
         single-document insert into the database
@@ -355,7 +464,7 @@ class Collection(object):
                 logger.error("Unable to complete transation (check the _id value for uniqueness). Document: %s", doc)
                 raise integrity_err
 
-    def insert(self, docs: list[dict], *args, **kwargs) -> None:
+    def insert(self, docs: Iterable[dict], *args, **kwargs) -> None:
         """
         multi-document insert into the database
 
@@ -392,10 +501,32 @@ class Collection(object):
                     logger.error("Discovered non-unique id values: %s", discovered_non_unique_id)
                 raise integrity_err
 
-    def bulk_write(self, docs: list[dict], *args, **kwargs) -> sqlite3.Cursor:
-        doc_objs = [item._doc for item in docs]
-        self.insert(doc_objs)
-        return Cursor(len(doc_objs))
+    def bulk_write(
+        self, docs: Iterable[Union[dict, InsertOne, ReplaceOne, UpdateOne]], *args, **kwargs
+    ) -> Sqlite3BulkWriteResult:
+        """
+        "Overridden method" to mimic the structure of mongodb as our design followed a lot of
+        patterns dictated by mongodb initially. If we wish to support sqlite3 with the command line
+        tooling then we have to structure this method similarly.
+
+        Across our codebase we have several potential different operations that are normally
+        associated with pymongo that are used with bulk_write. These are `InsertOne`, 'ReplaceOne`,
+        and 'UpdateOne`. We have to ensure on the sqlite3 version of bulk_write to expect handling
+        documents of this type
+        """
+        raw_documents = []
+        for document in docs:
+            try:
+                raw_documents.append(document._doc)
+            except Exception as gen_exp:
+                breakpoint()
+                pass
+
+        try:
+            self.insert(raw_documents)
+        except sqlite3.IntegrityError as integrity_error:
+            raise Sqlite3BulkWriteError from integrity_error
+        return Sqlite3BulkWriteResult(raw_documents)
 
     def update_one(self, query, what, upsert=False):
         assert len(what) == 1 and (
@@ -478,8 +609,3 @@ class Collection(object):
     def __getstate__(self):
         self.__dict__.pop("db", None)
         return self.__dict__
-
-
-class Cursor(object):
-    def __init__(self, inserted_count):
-        self.inserted_count = inserted_count
